backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==46.0.3

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0

# Async & Background Tasks
celery==5.3.4
redis==5.0.1

# Existing Python Engine Dependencies
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
matplotlib==3.8.2
openpyxl==3.1.2

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1
watchdog==3.0.0
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
